Move reset to gltf
